pipeline/requirements.txt
confluent-kafka
redis
pika
minio
python-dotenv
requests
streamlit==1.24.1